add notation of gnu

+++ src/ap_file.c
@@ -1,11 +1,10 @@
-//
-//  ap_file.c
-//  ap_file_system
-//
-//  Created by HU XUKAI on 14/11/12.
-//  Copyright (c) 2014年 HU XUKAI.<goingonhxk@gmail.com>
-//
-
+/*
+ *   Copyright (c) 2015, HU XUKAI
+ *
+ *   This source code is released for free distribution under the terms of the
+ *   GNU General Public License.
+ *
+ */
 #include <stdio.h>
 #include <string.h>
 #include <errno.h>
@@ -14,26 +13,16 @@
 #include <ap_pthread.h>
 #include <bag.h>
 
-/*每当一个独立线程调用ap_open时都会产生新的file结构
- *以后能使用hash使得同一个文件对应一个fd（相对于不同线程来说）
- */
-
 static inline int get_mask_from_oflag(int flag)
 {
-    int mask = 0;
-    switch (flag) {
-        case O_RDONLY:
-            mask = MAY_READ;
-            break;
-        case O_WRONLY:
-            mask = MAY_WRITE;
-            break;
-        case O_RDWR:
-            mask = MAY_WRITE | MAY_READ;
-            break;
-        default:
-            return -1;
-    }
+    int mask = MAY_READ;
+    if (flag & O_RDWR)
+        mask = MAY_READ | MAY_WRITE;
+    else if (flag & O_RDONLY)
+        mask = MAY_READ;
+    else if (flag & O_WRONLY)
+        mask = MAY_WRITE;
+    
     return mask;
 }
 
@@ -48,10 +37,6 @@ int ap_open(const char *path, int flags)
     }
     
     int mask = get_mask_from_oflag(flags);
-    if (mask == -1) {
-        errno = EINVAL;
-        return -1;
-    }
     
     SHOW_TRASH_BAG;
     struct ap_file *file;
@@ -161,11 +146,10 @@ static int __ap_mount(void *m_info, struct ap_file_system_type *fsyst, const cha
         return -1;
     }
     
-    if (!get) {
+    if (!get)
         parent = par_indic->cur_inode->parent;
-    }else{
+    else
         parent = par_indic->cur_inode;
-    }
     
     pthread_mutex_lock(&parent->ch_lock);
     if (!get){
@@ -188,9 +172,8 @@ static int __ap_mount(void *m_info, struct ap_file_system_type *fsyst, const cha
     
     if (mount_inode->name == NULL) {
         dir_name = strrchr(path, '/');
-        if (strcmp(path, "/")) {
+        if (strcmp(path, "/"))
             dir_name++;
-        }
         len = strlen(dir_name);
         mount_inode->name = Malloc_z(len + 1);
         strncpy(mount_inode->name, dir_name, len);
@@ -295,9 +278,6 @@ int ap_unmount(const char *path)
     B_return(de);
 }
 
-/*任何一个线程调用close都会立即关闭描述符释放file结构
- *这样其它的线程用同一fd就访问不到此文件了
- */
 int ap_close(int fd)
 {
     if (fd < 0 || fd > _OPEN_MAX) {
@@ -319,9 +299,8 @@ int ap_close(int fd)
     pthread_mutex_unlock(&file->file_lock);
     pthread_mutex_unlock(&file_info.files_lock);
     
-    if (file->f_ops->release != NULL) {
+    if (file->f_ops->release != NULL)
         file->f_ops->release(file, file->relate_i);
-    }
     AP_FILE_FREE(file);
     return 0;
 }
@@ -349,7 +328,7 @@ ssize_t ap_read(int fd, void *buf, size_t len)
     pthread_mutex_lock(&file->file_lock);
     pthread_mutex_unlock(&file_info.files_lock);
     
-    if (file->f_ops->read == NULL) {
+    if (file->f_ops == NULL ||  file->f_ops->read == NULL) {
         pthread_mutex_unlock(&file->file_lock);
         errno = ESRCH;
         return -1;
@@ -383,7 +362,8 @@ ssize_t ap_write(int fd, void *buf, size_t len)
     file = file_info.file_list[fd];
     pthread_mutex_lock(&file->file_lock);
     pthread_mutex_unlock(&file_info.files_lock);
-    if (file->f_ops->write == NULL) {
+    if (file->f_ops == NULL || file->f_ops->write == NULL) {
+        pthread_mutex_unlock(&file->file_lock);
         errno = ESRCH;
         return -1;
     }
@@ -404,7 +384,7 @@ off_t ap_lseek(int fd, off_t ops, int origin)
     struct ap_file *file;
     
     pthread_mutex_lock(&file_info.files_lock);
-    if (file_info.file_list[fd] == NULL) {
+    if (file->f_ops == NULL || file_info.file_list[fd] == NULL) {
         errno = EBADF;
         return -1;
     }
@@ -412,6 +392,7 @@ off_t ap_lseek(int fd, off_t ops, int origin)
     pthread_mutex_lock(&file->file_lock);
     pthread_mutex_unlock(&file_info.files_lock);
     if (file->f_ops->llseek == NULL) {
+        pthread_mutex_unlock(&file->file_lock);
         errno = ESRCH;
         return -1;
     }
@@ -475,9 +456,8 @@ int ap_mkdir(char *path, unsigned long mode)
     
     s_make = par_indic->cur_inode->i_ops->mkdir(par_indic); //由此函数直接负责将ap_inode插入到链表中
     AP_INODE_INICATOR_FREE(par_indic);
-    if (s_make == -1) {
+    if (s_make == -1)
         return -1;
-    }
     return 0;
 }
 
@@ -542,7 +522,7 @@ int ap_unlink(const char *path)
     int o = 0;
     
     if (link == 0) {     //已经没有其它目录链接此文件
-        if (op_inode->parent->i_ops->unlink != NULL) {
+        if (op_inode->i_ops != NULL && op_inode->i_ops->unlink != NULL) {
             int unlik_s = op_inode->parent->i_ops->unlink(op_inode);
             if (unlik_s == -1) {
                 errno = EBUSY;
@@ -712,7 +692,7 @@ int ap_rmdir(const char *path)
         return -1;
     }
     pthread_mutex_unlock(&op_inode->inode_counter.counter_lock);
-    if (op_inode->i_ops->rmdir != NULL) {
+    if (op_inode->i_ops != NULL && op_inode->i_ops->rmdir != NULL) {
         rm_s = op_inode->i_ops->rmdir(final_indc);
         if (rm_s != 0) {
             pthread_mutex_unlock(&op_inode->ch_lock);
@@ -737,9 +717,9 @@ int ap_rmdir(const char *path)
     pthread_mutex_unlock(&op_inode->mount_inode->mt_ch_lock);
     op_inode = convert_to_real_ind(op_inode);
     
-    if (op_inode->mount_inode->real_inode == op_inode) {
+    if (op_inode->mount_inode->real_inode == op_inode)
         AP_INODE_FREE(op_inode->mount_inode);
-    }
+
     AP_INODE_INICATOR_FREE(final_indc);
     return 0;
 }
@@ -806,9 +786,9 @@ AP_DIR *ap_open_dir(const char *path)
     }
     
     int get = walk_path(indc);
-    if (get == -1) {
+    if (get == -1)
         B_return(NULL);
-    }
+
     if (!indc->cur_inode->is_dir) {
         errno = ENOTDIR;
         B_return(NULL);
@@ -860,9 +840,9 @@ int ap_closedir(AP_DIR *dir)
         errno = EINVAL;
         return -1;
     }
-    if (dir->dir_i->i_ops->closedir != NULL) {
+    if (dir->dir_i->i_ops->closedir != NULL)
         dir->dir_i->i_ops->closedir(dir->dir_i);
-    }
+
     AP_DIR_FREE(dir);
     return 0;
 }
